pyproject.toml
[project]
name = "evf"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dearpygui",
    "Pillow",
    "numpy",
    "scipy",
    "tetra3",
    "playsound3",
]

[dependency-groups]
dev = [
    "nuitka>=4.0.2",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["python/evf"]

[tool.uv.sources]
tetra3 = { path = "python/vendor/tetra3", editable = true }
